requirements.txt
beautifulsoup4==4.9.3
brotlipy==0.7.0
certifi==2020.12.5
cffi==1.14.3
chardet==3.0.4
cheroot==8.4.8
cryptography==3.2.1
idna==2.10
jaraco.functools==3.0.1
lxml==4.6.2
more-itertools==8.6.0
numpy==1.19.4
opencv-contrib-python==4.4.0.46
opencv-python==4.4.0.46
pycparser==2.20
pyOpenSSL==19.1.0
pypinyin==0.40.0
PySocks==1.7.1
python-docx==0.8.10
requests==2.24.0
Rx==3.1.1
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.11
web.py==0.62
win-inet-pton==1.1.0
wincertstore==0.2
